scripts/requirements.txt
pyyaml
ffmpy